Solved CSES Bit_Strings

+++ CSES/Task/Bit_Strings/Bit_Strings.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main() {
   int n;
   cin >> n;
-  int MOD = 1e9 + 7;
+  long long MOD = 1e9 + 7;
   long long cal = pow(2, n);
   cout << cal % MOD;
 }
